[iOS] 별점 뷰 공통 구현으로 변경
- ReviewLabel 클래스 공통으로 재사용

issue: #109

+++ iOS/AirbnbApp/AirbnbApp/StayDetail/Models/StayDetail.swift
@@ -23,7 +23,7 @@ extension StayDetail {
     }
 
     struct ReviewInfo: Decodable {
-        let average: Float
+        let average: Double
         let numberOfReviews: Int
     }
